requirements.txt
GitPython
print_table